Move Auth and User modules to src/modules folder
